% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate_conda.R
\name{activate_conda}
\alias{activate_conda}
\title{Activate conda environment}
\usage{
activate_conda(
  conda_env = Sys.getenv("CONDA_ENV"),
  method = "reticulate",
  verbose = TRUE
)
}
\arguments{
\item{conda_env}{Conda environment to use.}

\item{method}{Method to use:
\itemize{
\item{"basilisk"}
\item{"reticulate"}
}}

\item{verbose}{Print messages.}
}
\value{
Null
}
\description{
Activate a conda environment using a global variable.
}
\examples{
activate_conda()
}
