% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anndata_to_sce.R
\name{anndata_to_sce}
\alias{anndata_to_sce}
\title{Convert: \code{Anndata} ==> \code{SingleCellExperiment}}
\usage{
anndata_to_sce(object, verbose = T)
}
\description{
Convert: \code{Anndata} ==> \code{SingleCellExperiment}
}
\examples{
library(scKirby)
data("example_anndata")
sce <- anndata_to_sce(example_anndata)
}
