% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anndata_to_se.R
\name{anndata_to_se}
\alias{anndata_to_se}
\title{Convert: \code{Anndata} ==> \code{SummarizedExperiment}}
\usage{
anndata_to_se(obj, as_sce = FALSE, verbose = TRUE)
}
\description{
Convert: \code{Anndata} ==> \code{SummarizedExperiment}
}
\examples{
obj <- example_obj("anndata")
sce <- anndata_to_se(obj)
}
