% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cor.R
\name{calc_cor}
\alias{calc_cor}
\title{Compute all pairwise trait correlations}
\usage{
calc_cor(
  obj,
  keys = NULL,
  graph_key = NULL,
  assay = NULL,
  slot = NULL,
  transpose = FALSE,
  method = "pearson",
  fill_na = NULL,
  use = "all.obs",
  nThreads = 1,
  verbose = TRUE
)
}
\arguments{
\item{obj}{Object of class:
\itemize{
\item{\code{DimReduc}}
\item{\code{Seurat}}
\item{\code{AnnData}}
\item{\code{prcomp}}
\item{\code{CellDataSet}}
\item{\code{list}}
\item{\code{matrix/data.frame}}
}}

\item{keys}{The keys of reductions to extract from.}

\item{method}{a character string indicating which correlation
    coefficient (or covariance) is to be computed.  One of
    \code{"pearson"} (default), \code{"kendall"}, or \code{"spearman"}:
    can be abbreviated.}

\item{use}{an optional character string giving a
    method for computing covariances in the presence
    of missing values.  This must be (an abbreviation of) one of the strings
    \code{"everything"}, \code{"all.obs"}, \code{"complete.obs"},
    \code{"na.or.complete"}, or \code{"pairwise.complete.obs"}.}

\item{nThreads}{ non-negative integer specifying the number of parallel threads to be used by certain
parts of correlation calculations. 
This option only has an effect on systems on which a POSIX thread
library is available (which currently includes Linux and Mac OSX, but excludes Windows).
If zero, the number of online processors will be used if it can be determined dynamically, otherwise
correlation calculations will use 2 threads. 
Note that this option does not affect what is usually the most expensive part of the
calculation, namely the matrix multiplication. The matrix multiplication is carried out by BLAS routines provided by R;
these can be sped up by installing a fast BLAS and making R use it. }

\item{verbose}{Print messages.}
}
\value{
A single-cell object with a graph.
}
\description{
Computes pairwise correlations for all traits
 using either \link[WGCNA]{cor} (faster) or
\link[stats]{cor} (slower).
}
\keyword{internal}
