% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_to_anndata.R
\name{cds_to_anndata}
\alias{cds_to_anndata}
\title{Convert: \code{CellDataSet} ==> \code{AnnData}}
\usage{
cds_to_anndata(
  obj,
  reimport = TRUE,
  save_path = tempfile(fileext = ".h5ad"),
  verbose = TRUE
)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{reimport}{Save and re-import the \code{AnnData} object into R to ensure
all data has been converted from Python-native to R-native objects
(e.g. pandas data.frames vs. R data.frames).}

\item{save_path}{Path to save the \code{H5Seurat} to.}

\item{verbose}{Print messages.}
}
\description{
Convert: \code{CellDataSet} ==> \code{AnnData}
}
\examples{
obj <- example_obj("cds")
obj2 <- cds_to_anndata(obj)
}
