% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_to_list.R
\name{cds_to_list}
\alias{cds_to_list}
\title{Convert: \code{CellDataSet} ==> \code{list}}
\usage{
cds_to_list(obj, verbose = TRUE)
}
\description{
Convert: \code{CellDataSet} ==> \code{list}
}
\examples{
obj <- example_obj("cds")
obj2 <- cds_to_list(obj)
}
