% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_to_sce.R
\name{cds_to_sce}
\alias{cds_to_sce}
\title{Convert: \code{CellDataSet} ==> \code{SingleCellExperiment}}
\usage{
cds_to_sce(object, verbose = T, as_sparse = T, as_DelayedArray = F)
}
\description{
Convert: \code{CellDataSet} ==> \code{SingleCellExperiment}
}
\examples{
library(scKirby)
data("example_cds")
sce <- cds_to_sce(object=example_cds)
}
