% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_to_se.R
\name{cds_to_se}
\alias{cds_to_se}
\title{Convert: \code{CellDataSet} ==> \code{SummarizedExperiment}}
\usage{
cds_to_se(
  obj,
  as_sce = FALSE,
  as_sparse = TRUE,
  as_delayedarray = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{as_sce}{Convert to a \link[SingleCellExperiment]{SingleCellExperiment}
format.}

\item{as_sparse}{Convert to a \link[Matrix]{sparseMatrix}.}

\item{as_delayedarray}{Convert to \link[DelayedArray]{DelayedArray}.}

\item{verbose}{Print messages.}
}
\description{
Convert: \code{CellDataSet} ==> \code{SummarizedExperiment}
}
\examples{
obj <- example_obj("cds")
obj2 <- cds_to_se(obj)
}
