% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_to_seurat.R
\name{cds_to_seurat}
\alias{cds_to_seurat}
\title{Convert: \code{CellDataSet} ==> \code{Seurat}}
\usage{
cds_to_seurat(object, verbose = T)
}
\description{
Convert: \code{CellDataSet} ==> \code{Seurat}
}
\examples{
library(scKirby)
data("example_cds")

seurat <- cds_to_seurat(example_cds)
}
