% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dict.R
\name{class_dict}
\alias{class_dict}
\title{Class dictionary}
\usage{
class_dict(group = NULL)
}
\arguments{
\item{group}{Name of one or more single-cell group of
single-cell object classes.}
}
\value{
Named list of object classes.
}
\description{
Single-cell object class dictionary.
}
\examples{
classes <- class_dict()
}
