% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R, R/to_seurat.R
\name{converters}
\alias{converters}
\alias{to_seurat}
\title{Converters}
\usage{
converters()

to_seurat(
  obj,
  as_h5seurat = FALSE,
  update = TRUE,
  save_path = tempfile(fileext = ".h5seurat"),
  verbose = TRUE
)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{as_h5seurat}{Convert to the \code{H5Seurat} subclass.}

\item{update}{Ensure the object is updated to the latest version of Seurat.}

\item{save_path}{Path to save the \code{H5Seurat} to.}

\item{verbose}{Print messages.}
}
\description{
Functions to convert one single-cell format to another.
}
\section{Functions}{
\itemize{
\item \code{to_seurat()}: Convert any single-cell object to
\code{Seurat} format.

}}
\examples{
map_dt <- converters()
obj <- example_obj("cds")
obj2 <- to_seurat(obj)
}
