% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd_to_se.R
\name{ctd_to_se}
\alias{ctd_to_se}
\title{Convert: \code{CellTypeDataset} ==> \code{SingleCellExperiment}}
\usage{
ctd_to_se(obj, as_sce = FALSE, as_delayedarray = FALSE, verbose = TRUE)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{as_sce}{Convert to a \link[SingleCellExperiment]{SingleCellExperiment}
format.}

\item{as_delayedarray}{Convert to \link[DelayedArray]{DelayedArray}.}

\item{verbose}{Print messages.}
}
\description{
Convert: \code{CellTypeDataset} ==> \code{SingleCellExperiment}
}
\examples{
obj <- example_obj("ctd")
obj <- ctd_to_sce(obj)
}
