% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd_to_seurat.R
\name{ctd_to_seurat}
\alias{ctd_to_seurat}
\title{Convert: \code{CellTypeDataSet} ==> \code{Seurat}}
\usage{
ctd_to_seurat(obj, verbose = TRUE)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{verbose}{Print messages.}
}
\description{
Convert: \code{CellTypeDataSet} ==> \code{Seurat}
}
\examples{
obj <- example_obj("ctd")
obj2 <- ctd_to_seurat(obj)
}
