% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimreduc_to_list.R
\name{dimreduc_to_list}
\alias{dimreduc_to_list}
\title{Convert: \code{AnnData} ==> \code{list}}
\usage{
dimreduc_to_list(obj)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}
}
\description{
Convert: \code{AnnData} ==> \code{list}
}
\examples{
obj <- example_obj("anndata")
obj2 <- anndata_to_list(obj)
}
