% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWCEctd_to_sce.R
\name{EWCEctd_to_sce}
\alias{EWCEctd_to_sce}
\title{Convert: \code{EWCEctd} ==> \code{SingleCellExperiment}}
\usage{
EWCEctd_to_sce(object, verbose = T)
}
\description{
Convert: \code{EWCEctd} ==> \code{SingleCellExperiment}
}
\examples{
library(scKirby)
data("example_ctd")
sce <- EWCEctd_to_sce(example_ctd)
}
