% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWCEctd_to_seurat.R
\name{EWCEctd_to_seurat}
\alias{EWCEctd_to_seurat}
\title{Convert: \code{EWCEctd} ==> \code{Seurat}}
\usage{
EWCEctd_to_seurat(object, verbose = T)
}
\description{
Convert: \code{EWCEctd} ==> \code{Seurat}
}
\examples{
library(scKirby)
data("example_ctd")
seurat <- EWCEctd_to_seurat(example_ctd)
}
