% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWCElist_to_sce.R
\name{EWCElist_to_sce}
\alias{EWCElist_to_sce}
\title{Convert: \code{EWCElist} ==> \code{SingleCellExperiment}}
\usage{
EWCElist_to_sce(object, verbose = T)
}
\description{
Convert: \code{EWCElist} ==> \code{SingleCellExperiment}
}
\examples{
library(scKirby)
data("example_EWCElist")
sce <- EWCElist_to_sce(example_EWCElist)
}
