% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWCElist_to_seurat.R
\name{EWCElist_to_seurat}
\alias{EWCElist_to_seurat}
\title{Convert: \code{EWCElist} ==> \code{Seurat}}
\usage{
EWCElist_to_seurat(object, verbose = T)
}
\description{
Convert: \code{EWCElist} ==> \code{Seurat}
}
\examples{
library(scKirby)
data("example_EWCElist")
seurat <- EWCElist_to_seurat(object=example_EWCElist)
}
