% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_cds}
\alias{example_cds}
\title{Example \code{CellDataSet}}
\format{
An object of class \code{CellDataSet} with 230 rows and 80 columns.
}
\usage{
example_cds
}
\description{
Example \code{CellDataSet}
}
\examples{
\dontrun{
library(Seurat)
data("pbmc_small")
example_seurat <- Seurat::UpdateSeuratObject(pbmc_small)
example_cds <- Seurat::as.CellDataSet(example_seurat)
usethis::use_data(example_cds, overwrite = T)
}
}
\keyword{datasets}
