% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_ctd}
\alias{example_ctd}
\title{Example \code{CellTypeDataset}}
\format{
An object of class \code{list} of length 2.
}
\usage{
example_ctd
}
\description{
Example \code{CellTypeDataset}
}
\examples{
\dontrun{
example_ctd <- ewceData::ctd()
### Remove dendrograms to reduce file size
example_ctd[[1]]$plotting <- NULL
example_ctd[[2]]$plotting <- NULL
usethis::use_data(example_ctd, overwrite = T)
}
}
\keyword{datasets}
