% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_EWCElist}
\alias{example_EWCElist}
\title{Example \code{EWCElist}}
\format{
An object of class \code{list} of length 2.
}
\usage{
example_EWCElist
}
\description{
Example \code{EWCElist}
}
\examples{
\dontrun{
example_EWCElist <- ewceData::cortex_mrna()
usethis::use_data(example_EWCElist, overwrite = T)
}
}
\keyword{datasets}
