% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_sce}
\alias{example_sce}
\title{Subsetted mouse ESC \code{SingleCellExperiment}  object}
\format{
A \code{SingleCellExperiment} object
}
\source{
\url{https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-2600/}
}
\usage{
example_sce
}
\description{
A dataset containing 300 cells and 2026 genes from two batches of mouse ESC data.
}
\details{
Copied from the R package \pkg{scMerge}.
}
\examples{
\dontrun{
if(!require(scMerge))BiocManager::install("scMerge")
data('example_sce', package = 'scMerge')
batch_names<-unique(example_sce$batch)
usethis::use_data(example_sce, overwrite = T)
}
}
\references{
Kolodziejczyk et al.
}
\keyword{datasets}
