% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_seurat}
\alias{example_seurat}
\title{Example \code{Seurat}}
\format{
An object of class \code{Seurat} with 230 rows and 80 columns.
}
\usage{
example_seurat
}
\description{
Example \code{Seurat}
}
\examples{
\dontrun{
library(Seurat)
data("pbmc_small")
example_seurat <- Seurat::UpdateSeuratObject(pbmc_small)
usethis::use_data(example_seurat, overwrite = T)
}
}
\keyword{datasets}
