% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cor.R
\name{get_cor}
\alias{get_cor}
\title{Extract correlation matrix}
\usage{
get_cor(
  obj,
  keys = NULL,
  assay = NULL,
  slot = NULL,
  graph_key = NULL,
  method = "pearson",
  return_obj = FALSE,
  as_graph = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{keys}{The keys of reductions to extract from.}

\item{assay}{Assay to used to compute correlation graph
if one does not already exist.}

\item{slot}{Name of component object to return}

\item{graph_key}{Name of the graph to extract.}

\item{method}{Pairwise correlation method.}

\item{return_obj}{Whether to return the single-cell object with a new
\code{graph}, or to simply return the sparse correlation matrix.}

\item{as_graph}{Convert the correlation matrix to the
\link[SeuratObject]{Graphs} class.}

\item{verbose}{Print messages.}
}
\value{
Trait-trait correlation matrix.
}
\description{
Extract correlation matrix stored as a graph.
}
\examples{
obj <- example_obj("seurat")
Xcor <- get_cor(obj = obj,
                keys = "pca")
}
