% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data}
\usage{
get_data(
  obj,
  transpose = TRUE,
  simplify = FALSE,
  as_sparse = FALSE,
  verbose = TRUE
)
}
\description{
Extract expression matrix metadata from any single-cell object.
}
\examples{
obj <- example_obj("loom")
X <- get_data(obj)
}
