% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_graphs.R
\name{get_graphs}
\alias{get_graphs}
\title{Get graphs}
\usage{
get_graphs(obj, keys = NULL, n = NULL, verbose = TRUE)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{keys}{The keys of reductions to extract from.}

\item{n}{The maximum number of elements to return from that list.}

\item{verbose}{Print messages.}
}
\value{
Named list of graph objects.
}
\description{
Get graph objects from any single-cell object class.
}
\examples{
obj <- example_obj("seurat")
g <- get_graphs(obj)
}
