% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_elements.R
\name{get_n_elements}
\alias{get_n_elements}
\title{Get N elements}
\usage{
get_n_elements(l, n = 1, simplify = TRUE, verbose = TRUE)
}
\arguments{
\item{l}{A list.}

\item{n}{The maximum number of elements to return from that list.}

\item{simplify}{When the final length of the list is 1, unlist the object.}

\item{verbose}{Print messages.}
}
\value{
A shortened list or a single element.
}
\description{
Get only the first N elements from a named list.
}
\examples{
l <- lapply(seq(4),prod,3)
l1 <- get_n_elements(l, n=1)
l2 <- get_n_elements(l, n=2, simplify = FALSE)
}
