% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obs.R
\name{get_obs}
\alias{get_obs}
\title{Get observations metadata}
\usage{
get_obs(obj, rownames_col = NULL, verbose = TRUE)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{rownames_col}{Name of the column to use as row names in the metadata
(i.e. unique cell IDs/barcodes).}

\item{verbose}{Print messages.}
}
\value{
An observation (sample) metadata data.frame.
}
\description{
Extract sample observations (i.e. cell) metadata from any single-cell object.
}
\examples{
obj <- example_obj("scle")
obs <- get_obs(obj)
}
