% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obs_names.R
\name{get_obs_names}
\alias{get_obs_names}
\title{Get obs names}
\usage{
get_obs_names(obj)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}
}
\value{
Vector of observation names.
}
\description{
Get the names of each observation in a single-cell object.
}
\examples{
obj <- example_obj()
onames <- get_obs_names(obj)
}
