% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obsm.R
\name{get_obsm}
\alias{get_obsm}
\title{Extract reductions}
\usage{
get_obsm(obj, keys = NULL, n = NULL, verbose = TRUE)
}
\arguments{
\item{obj}{Object of class:
\itemize{
\item{\code{DimReduc}}
\item{\code{Seurat}}
\item{\code{AnnData}}
\item{\code{prcomp}}
\item{\code{CellDataSet}}
\item{\code{list}}
\item{\code{matrix/data.frame}}
}}

\item{keys}{The keys of reductions to extract from.}

\item{n}{The maximum number of elements to return from that list.}

\item{verbose}{Print messages.}
}
\value{
A matrix of cell embeddings coordinates.
}
\description{
Extract dimensionality keys objects from any single-cell object class.
}
\examples{
obj <- example_obj("cds")
obsm <- get_obsm(obj)
}
