% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reductions.R
\name{get_reductions}
\alias{get_reductions}
\title{Extract reductions}
\usage{
get_reductions(obj, keys = NULL, verbose = TRUE)
}
\arguments{
\item{obj}{Object of class:
\itemize{
\item{\code{DimReduc}}
\item{\code{Seurat}}
\item{\code{Seurat}}
}}

\item{keys}{The keys of reductions to extract from.}

\item{verbose}{Print messages.}
}
\value{
A named list with up to three elements:
\itemize{
\item{"embeddings": } Embeddings of samples (cells) in low-dimensional space.
\item{"loadings": } Loadings of features (genes) for each latent component.
\item{"loadings_projected": } Projected loadings of features (genes)
 for each latent component.
}
}
\description{
Extract dimensionality keys objects from any single-cell object class.
}
\examples{
obj <- example_obj("ad")
red <- get_reductions(obj)
}
