% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uns.R
\name{get_uns}
\alias{get_uns}
\title{Get unstructured metadata}
\usage{
get_uns(obj, verbose = TRUE)
}
\description{
Extract unstructuerd metadata from any single-cell object.
}
\examples{
obj <- example_obj("s")
uns <- get_uns(obj)
}
