% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var.R
\name{get_var}
\alias{get_var}
\title{Get variable metadata}
\usage{
get_var(obj, rownames_col = NULL, simplify = FALSE, verbose = TRUE)
}
\arguments{
\item{rownames_col}{Name of the column to use as row names in the metadata
(i.e. unique gene/transcript IDs).}
}
\description{
Extract feature variable (i.e. gene) metadata from any single-cell object.
}
\examples{
obj <- example_obj("cds")
var <- get_var(obj)
}
