% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_names.R
\name{get_var_names}
\alias{get_var_names}
\title{Get var names}
\usage{
get_var_names(obj)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}
}
\value{
Vector of observation names.
}
\description{
Get the names of each feature in a single-cell object.
}
\examples{
obj <- example_obj()
vnames <- get_obs_names(obj)
}
