% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variable_features.R
\name{get_variable_features}
\alias{get_variable_features}
\title{Get variable features}
\usage{
get_variable_features(obj, reduce = NULL)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{reduce}{Function with which to reduce variable features across assays
into a single vector (e.g. \code{intersect} or \code{union}).}
}
\value{
A named list with variable features for each assay.
}
\description{
Get pre-computed variable features (e.g. genes)
stored within single-cell objects.
}
\examples{
obj <- example_obj("seurat")
feat <- get_variable_features(obj)
}
