% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_varm.R
\name{get_varm}
\alias{get_varm}
\title{Extract variable loadings}
\usage{
get_varm(obj, keys = NULL, n = NULL, verbose = TRUE)
}
\arguments{
\item{obj}{Object of class:
\itemize{
\item{\code{DimReduc}}
\item{\code{Seurat}}
\item{\code{AnnData}}
\item{\code{prcomp}}
\item{\code{CellDataSet}}
\item{\code{list}}
\item{\code{matrix/data.frame}}
}}

\item{keys}{The keys of reductions to extract from.}

\item{n}{The maximum number of elements to return from that list.}

\item{verbose}{Print messages.}
}
\value{
A named list of matrices containing the feature loadings.
}
\description{
Extract variable feature loadings from any single-cell object class.
}
\examples{
obj <- example_obj("seurat")
varm <- get_varm(obj)
}
