% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_x.R
\name{get_x}
\alias{get_x}
\title{Get data}
\usage{
get_x(
  obj,
  transpose = TRUE,
  n = NULL,
  assay = NULL,
  slot = NULL,
  as_sparse = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{transpose}{Transpose the matrix with \code{Matrix::t}.}

\item{n}{The maximum number of elements to return from that list.}

\item{assay}{Name of the initial assay}

\item{slot}{Name of component object to return}

\item{as_sparse}{Convert to a \link[Matrix]{sparseMatrix}.}

\item{verbose}{Print messages.}
}
\value{
A named list of matrices.
}
\description{
Extract expression matrix metadata from any single-cell object.
}
\examples{
obj <- example_obj("seurat")
X <- get_x(obj)
}
