% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5seurat_to_anndata.R
\name{h5seurat_to_anndata}
\alias{h5seurat_to_anndata}
\title{Convert: \code{h5Seurat} ==> \code{anndata}}
\usage{
h5seurat_to_anndata(
  obj,
  reimport = TRUE,
  save_path = tempfile(fileext = ".h5ad"),
  overwrite = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{reimport}{Save and re-import the \code{AnnData} object into R to ensure
all data has been converted from Python-native to R-native objects
(e.g. pandas data.frames vs. R data.frames).}

\item{save_path}{Path to save the \code{H5Seurat} to.}

\item{overwrite}{Overwrite existing \code{dest}}

\item{verbose}{Print messages.}

\item{...}{
  Arguments passed on to \code{\link[anndata:write_h5ad]{anndata::write_h5ad}}
  \describe{
    \item{\code{anndata}}{An \code{\link[anndata:AnnData]{AnnData()}} object}
    \item{\code{filename}}{Filename of data file. Defaults to backing file.}
    \item{\code{compression}}{See the h5py \href{http://docs.h5py.org/en/latest/high/dataset.html#dataset-compression}{filter pipeline}.
Options are \code{"gzip"}, \code{"lzf"} or \code{NULL}.}
    \item{\code{compression_opts}}{See the h5py \href{http://docs.h5py.org/en/latest/high/dataset.html#dataset-compression}{filter pipeline}.}
    \item{\code{as_dense}}{Sparse in AnnData object to write as dense. Currently only supports \code{"X"} and \code{"raw/X"}.}
  }}
}
\description{
Convert: \code{h5Seurat} ==> \code{anndata}
}
\examples{
obj <- example_obj("h5seurat")
obj2 <- h5seurat_to_anndata(obj)
}
