% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5seurat_to_seurat.R
\name{h5seurat_to_seurat}
\alias{h5seurat_to_seurat}
\title{Convert: \code{Seurat} ==> \code{H5Seurat}}
\usage{
h5seurat_to_seurat(obj, verbose = TRUE, ...)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{verbose}{Print messages.}

\item{...}{
  Arguments passed on to \code{\link[SeuratObject:as.Seurat]{SeuratObject::as.Seurat}}
  \describe{
    \item{\code{x}}{An object to convert to class \code{Seurat}}
  }}
}
\description{
Convert: \code{Seurat} ==> \code{H5Seurat}
}
\examples{
obj <- example_obj("h5seurat")
seurat <- h5seurat_to_seurat(obj)
}
