% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_graph.R
\name{has_graph}
\alias{has_graph}
\title{Has graph}
\usage{
has_graph(obj, graph_keys, verbose = TRUE)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{graph_keys}{The names of specific graphs to extract.}

\item{verbose}{Print messages.}
}
\description{
Check whether an object has a particular graph.
}
