% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_data.R
\name{ingest_data}
\alias{ingest_data}
\title{Import and standardize scRNAseq data across different formats}
\source{
\href{https://mojaveazure.github.io/seurat-disk/articles/convert-anndata.html}{SeuratDisk}
\href{https://github.com/rcannood/anndata}{anndata (R)}
\href{https://anndata.readthedocs.io/en/latest/}{anndata (python)}
\href{https://satijalab.org/loomR/loomR_tutorial.html}{loomR}
\href{https://bioconductor.org/packages/release/bioc/vignettes/SingleCellExperiment/inst/doc/intro.html}{SingleCellExperiment}
\href{https://petehaitch.github.io/BioC2020_DelayedArray_workshop/articles/Effectively_using_the_DelayedArray_framework_for_users.html}{DelayedArray workshop}
\href{https://theislab.github.io/zellkonverter/articles/zellkonverter.html}{zellkonverter}
}
\usage{
ingest_data(
  obj,
  input_type = "guess",
  output_class = c("SingleCellExperiment", "Seurat", "CellDataSet", "list"),
  custom_reader = NULL,
  save_path = NULL,
  overwrite = FALSE,
  return_save_path = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{obj}{A single-cell data object, or a path to saved single-cell data.}

\item{input_type}{Format of \code{obj}. By default, the type will be inferred.}

\item{output_class}{Format to convert \code{obj} to.}

\item{custom_reader}{Custom function to read \code{obj} into R.}

\item{save_path}{Path to save the converted \code{obj}.}

\item{overwrite}{If a file of the same name exists, overwrite it.}

\item{return_save_path}{If \code{TRUE}, a list with both the converted object
and the saved file path will be returned (instead of just the converted object).}

\item{verbose}{Print messages.}

\item{...}{
  Arguments passed on to \code{\link[=read_data]{read_data}}
  \describe{
    \item{\code{path}}{Path to saved single-cell data file,
or a directory containing the data files.
If \code{path} is already an R object, it will be returned directly.}
    \item{\code{filetype}}{The type of file trying to be read in,
made explicit by the user.
"guess" (default) will simply infer the most likely file type.}
    \item{\code{as_sparse}}{Convert to a \link[Matrix]{sparseMatrix}.}
  }}
}
\value{
Converted single-cell object.
}
\description{
Automatically infers data format of scRNAseq object, or a path to that object.
It then uses the appropriate functions to import that data and convert it to a
\pkg{SingleCellExperiment}, which is recognized by other \pkg{EWCE} functions.
}
\examples{
\dontrun{
#### Input: Seurat object ####
seurat <- example_obj("seurat")
sce1 <- ingest_data(obj=seurat)

#### Input: Matrix object ####
sce2 <- ingest_data(obj=seurat@assays$RNA@counts)

#### Input hdf5se object ####
hdf5se <- example_obj("hdf5se")
sce3 <- ingest_data(obj=hdf5se)

#### Input anndata object ####
ad <- example_obj("anndata")
sce4 <- ingest_data(obj=ad)

#### Input anndata path ####
sce5 <- ingest_data(obj = ad$filename)

#### Input: h5Seurat object ####
h5seurat <- example_obj("h5seurat")
sce6 <- ingest_data(obj=h5seurat)

#### Input: h5Seurat path ####
sce7 <- ingest_data(obj=h5seurat$path)

#### Input: loom object ####
library(Seurat)
loom <- example_obj("loom")
sce8 <- ingest_data(obj=loom)

#### Input: EWCE object ####
\dontrun{
ewce <- ewceData::cortex_mrna()
sce10 <- ingest_data(obj=ewce)
}
}
}
