% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_class.R
\name{is_class}
\alias{is_class}
\title{Is class}
\usage{
is_class(obj, group = NULL)
}
\arguments{
\item{obj}{Data object.}

\item{group}{Name of one or more single-cell group of
single-cell object classes.}
}
\description{
When \code{group} is not supplied (default),
determines which group of single-cell classes the object belongs to (if any).
When \code{group} is supplied,
determine whether a single-cell object belongs to a particular group of
object classes.
See \link[scKirby]{class_dict} for details.
}
\examples{
obj <- example_obj("Seurat")
is_class(obj,"seurat")
}
