% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_installed.R
\name{is_installed}
\alias{is_installed}
\title{Is package installed}
\usage{
is_installed(pkg, error = FALSE)
}
\arguments{
\item{pkg}{Name of an R package to check.}

\item{error}{If \code{TRUE}, produces an error if the \code{pkg}
isn't installed.
If \code{FALSE}, produces a warning if the package isn't installed.}
}
\value{
Boolean
}
\description{
Is package installed
}
\keyword{internal}
