% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_anndata.R
\name{list_to_anndata}
\alias{list_to_anndata}
\title{Convert: \code{list} ==> \code{AnnData}}
\usage{
list_to_anndata(obj, transpose = TRUE, verbose = TRUE)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{transpose}{Transpose the matrix with \code{Matrix::t}.}

\item{verbose}{Print messages.}
}
\description{
Convert: \code{list} ==> \code{AnnData}
}
\examples{
obj <- example_obj("list")
obj2 <- list_to_anndata(obj)
}
