% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_dimreduc.R
\name{list_to_dimreduc}
\alias{list_to_dimreduc}
\title{Convert: \link[base]{list} ==> \code{DimReducObject}}
\usage{
list_to_dimreduc(obj, assay, key, ...)
}
\arguments{
\item{obj}{A named list containing:
\itemize{
\item{obsm : }{Sample embeddings}
\item{varm : }{Feature loadings}
\item{stdev : Standard deviation (if applicable) for the
 dimensional reduction}
}}

\item{assay}{Assay to use.}

\item{key}{Key to use (name of embedding).}

\item{...}{
  Arguments passed on to \code{\link[SeuratObject:CreateDimReducObject]{SeuratObject::CreateDimReducObject}}
  \describe{
    \item{\code{embeddings}}{A matrix with the cell embeddings}
    \item{\code{loadings}}{A matrix with the feature loadings}
    \item{\code{projected}}{A matrix with the projected feature loadings}
    \item{\code{stdev}}{Standard deviation (if applicable) for the dimensional reduction}
    \item{\code{global}}{Specify this as a global reduction (useful for visualizations)}
    \item{\code{jackstraw}}{Results from the JackStraw function}
    \item{\code{misc}}{list for the user to store any additional information associated
with the dimensional reduction}
  }}
}
\description{
Create a \pkg{Seurat} \code{DimReducObject}
from embeddings and loadings.
}
