% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_se.R
\name{list_to_se}
\alias{list_to_se}
\title{Convert: \code{list} ==> \code{SummarizedExperiment}}
\usage{
list_to_se(obj, as_sce = FALSE, verbose = TRUE)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{as_sce}{Convert to a \link[SingleCellExperiment]{SingleCellExperiment}
format.}

\item{verbose}{Print messages.}
}
\description{
Convert: \code{list} ==> \code{SummarizedExperiment}
}
\examples{
obj <- example_obj("list")
obj2 <- list_to_se(obj)
obj3 <- list_to_se(obj, as_sce=TRUE)
}
