% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_seurat.R
\name{list_to_seurat}
\alias{list_to_seurat}
\title{Convert: \code{list} ==> \code{Seurat}}
\usage{
list_to_seurat(obj, verbose = TRUE)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{verbose}{Print messages.}
}
\description{
Convert: \code{list} ==> \code{Seurat}
}
\examples{
obj <- example_obj("list")
obj2 <- list_to_seurat(obj)
}
