% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadRData}
\alias{loadRData}
\title{loadRData}
\usage{
loadRData(fileName)
}
\description{
Load processed data using a function that assigns it
to a specific variable (so you don't have to guess what the loaded variable name is).
}
