% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom_to_list.R
\name{loom_to_list}
\alias{loom_to_list}
\title{Convert: \code{loom} ==> \code{list}}
\usage{
loom_to_list(obj, verbose = TRUE)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{verbose}{Print messages.}
}
\description{
Convert: \code{loom} ==> \code{list}
}
\examples{
library(Seurat)
obj <- example_obj("loom")
obj2 <- loom_to_list(obj)
}
