% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom_to_sce.R
\name{loom_to_sce}
\alias{loom_to_sce}
\title{Convert: \code{loom} ==> \code{SingleCellExperiment}}
\usage{
loom_to_sce(object, verbose = T, ...)
}
\description{
Convert: \code{loom} ==> \code{SingleCellExperiment}
}
\examples{
library(scKirby)
sce <- loom_to_sce(object=example_loom())
}
