% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom_to_scle.R
\name{loom_to_scle}
\alias{loom_to_scle}
\title{Convert: \code{loom} ==> \code{SingleCellLoomExperiment}}
\usage{
loom_to_scle(obj, verbose = TRUE, ...)
}
\description{
Convert: \code{loom} ==> \code{SingleCellLoomExperiment}
}
\examples{
library(Seurat)
obj <-  example_obj("loom")
sce <- loom_to_scle(obj)
}
