% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom_to_se.R
\name{loom_to_se}
\alias{loom_to_se}
\title{Convert: \code{loom} ==> \code{SummarizedExperiment}}
\usage{
loom_to_se(obj, as_sce = FALSE, verbose = TRUE, ...)
}
\description{
Convert: \code{loom} ==> \code{SummarizedExperiment}
}
\examples{
library(Seurat)
obj <- example_obj("loom")
obj2 <- loom_to_se(obj)
}
