% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices_to_assayobjects.R
\name{matrices_to_assayobjects}
\alias{matrices_to_assayobjects}
\title{Matrices to AssayObjects}
\usage{
matrices_to_assayobjects(matrices, var_features = NULL, verbose = TRUE)
}
\arguments{
\item{matrices}{A named list of matrices.}

\item{meta.features}{[Optional] A named list of \code{meta.features}
(one per unique assay in \code{matrices}).}
}
\value{
a named list of \pkg{Seurat} assay objects.
}
\description{
Convert a named list of matrices to a named list of \pkg{Seurat}
assay objects (see \link[SeuratObject]{CreateAssayObject}).
}
\examples{
obj <- example_obj("seurat")
l <- seurat_to_list(obj)
assaylist <- matrices_to_assayobjects(matrices = l$data,
                                      var_features = l$var_features)
}
