% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_sce.R
\name{matrix_to_sce}
\alias{matrix_to_sce}
\title{Convert: \code{matrix} ==> \code{SingleCellExperiment}}
\usage{
matrix_to_sce(object, verbose = T)
}
\description{
Convert: \code{matrix} ==> \code{SingleCellExperiment}
}
\examples{
library(scKirby)
data("example_seurat")
sce <- matrix_to_sce(example_seurat@assays$RNA@counts)
}
