% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_se.R
\name{matrix_to_se}
\alias{matrix_to_se}
\title{Convert: \code{matrix} ==> \code{SummarizedExperiment}}
\usage{
matrix_to_se(obj, as_sce = FALSE, verbose = TRUE)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{as_sce}{Convert to a \link[SingleCellExperiment]{SingleCellExperiment}
format.}

\item{verbose}{Print messages.}
}
\description{
Convert: \code{matrix} ==> \code{SummarizedExperiment}
}
\examples{
obj <- example_obj("matrix")
obj2 <- matrix_to_se(obj)
}
