% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_seurat.R
\name{matrix_to_seurat}
\alias{matrix_to_seurat}
\title{Convert: \code{matrix} ==> \code{Seurat}}
\usage{
matrix_to_seurat(obj, verbose = TRUE)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{verbose}{Print messages.}
}
\description{
Convert: \code{matrix} ==> \code{Seurat}
}
\examples{
obj <- example_obj("matrix")
obj2 <- matrix_to_seurat(obj)
}
