% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_anndata.R
\name{merge_anndata}
\alias{merge_anndata}
\title{Merge \code{anndata} objects}
\usage{
merge_anndata(
  obj_list,
  axis = 0L,
  join = "outer",
  merge = NULL,
  uns_merge = NULL,
  label = NULL,
  keys = NULL,
  index_unique = NULL,
  fill_value = NULL,
  pairwise = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{obj_list}{A named list of \link[anndata]{AnnData} objects,
or paths to \link[anndata]{AnnData} objects
saved to disk as \emph{.h5ad} files.}

\item{axis}{Which axis to concatenate along.}

\item{join}{How to align values when concatenating. If "outer", the union of the other axis is taken. If "inner", the intersection. See \code{concatenation} for more.}

\item{merge}{How elements not aligned to the axis being concatenated along are selected. Currently implemented strategies include: * \code{NULL}: No elements are kept. * \code{"same"}: Elements that are the same in each of the objects. * \code{"unique"}: Elements for which there is only one possible value. * \code{"first"}: The first element seen at each from each position. * \code{"only"}: Elements that show up in only one of the objects.}

\item{uns_merge}{How the elements of \code{.uns} are selected. Uses the same set of strategies as the \code{merge} argument, except applied recursively.}

\item{label}{Column in axis annotation (i.e. \code{.obs} or \code{.var}) to place batch information in. If it's NULL, no column is added.}

\item{keys}{Names for each object being added. These values are used for column values for \code{label} or appended to the index if \code{index_unique} is not \code{NULL}. Defaults to incrementing integer labels.}

\item{index_unique}{Whether to make the index unique by using the keys. If provided, this is the delimeter between "{orig_idx}{index_unique}{key}". When \code{NULL}, the original indices are kept.}

\item{fill_value}{When \code{join="outer"}, this is the value that will be used to fill the introduced indices. By default, sparse arrays are padded with zeros, while dense arrays and DataFrames are padded with missing values.}

\item{pairwise}{Whether pairwise elements along the concatenated dimension should be included. This is FALSE by default, since the resulting arrays are often not meaningful.}

\item{verbose}{Print messages.}
}
\value{
A merged \link[anndata]{AnnData} object.
}
\description{
Merge a list of Seurat objects efficiently.
}
\examples{
obj <- example_obj("anndata")
obj_list <- Seurat::SplitObject(obj,split.by = "groups")
obj2 <- merge_seurat(obj_list)
}
