% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mofa_to_dro.R
\name{mofa_to_dro}
\alias{mofa_to_dro}
\title{Convert: \code{MOFA} ==> \code{DimReducObject}}
\usage{
mofa_to_dro(obj, keys = NULL, assay = NULL, verbose = TRUE)
}
\arguments{
\item{obj}{A trained MOFA model.}

\item{keys}{The keys of reductions to extract from.}

\item{assay}{Assay used to calculate this dimensional reduction}

\item{verbose}{Print messages.}
}
\value{
A named list of a \pkg{Seurat} \code{DimReducObject}.
}
\description{
Convert: \code{MOFA} ==> \code{DimReducObject}
}
