% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_dict.R
\name{output_dict}
\alias{output_dict}
\title{Output dictionary}
\usage{
output_dict(output_class = NULL)
}
\arguments{
\item{output_class}{Name of the output class you wish to produce.
Alternatively, set to \code{NULL} to return all possible output classes.}
}
\value{
A valid \code{output_class}.
}
\description{
A list of all object classes that the single-cell dataset can be returned as.
Classes are grouped by families of related classes.
}
\examples{
output_class <- output_dict(output_class="seurat")
}
