% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_metadata.R
\name{prepare_metadata}
\alias{prepare_metadata}
\title{Prepare metadata}
\usage{
prepare_metadata(
  obj,
  save_path = NULL,
  species,
  standardise_species = TRUE,
  method = "gprofiler",
  level_cols = list(),
  extra_metadata = NULL,
  return_metadata = FALSE,
  verbose = TRUE
)
}
\description{
Standardise metadata for any single-cell object.
}
\examples{
obj <- example_obj("anndata")
obj2 <- prepare_metadata()
}
