% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read data}
\usage{
read_data(path, filetype = "guess", custom_reader = NULL, verbose = TRUE, ...)
}
\arguments{
\item{path}{Path to saved single-cell data file,
or a directory containing the data files.
If \code{path} is already an R object, it will be returned directly.}

\item{filetype}{The type of file trying to be read in,
made explicit by the user.
"guess" (default) will simply infer the most likely file type.}

\item{custom_reader}{A user-supplied function to read in the data with.}

\item{verbose}{Print messages.}
}
\value{
An single-cell data object.
 The object class that the data gets imported as depends on file type.
}
\description{
Automatically infer the format of a single-cell dataset on disk and infer
how to read it into R.
}
\examples{
library(Seurat)
obj <- example_obj("loom")
obj2 <- read_data(obj$filename)
}
