% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data.R
\name{save_data}
\alias{save_data}
\title{Save data}
\usage{
save_data(
  obj,
  filetype = c("h5", "h5seurat", "h5ad", "rda", "rds", "anndata"),
  save_path = file.path(tempdir(), paste("scKirby_output", filetype, sep = ".")),
  quicksave_hdf5 = TRUE,
  overwrite = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{obj}{A single-cell data object, or a path to saved single-cell data.}

\item{filetype}{The file format to write the object as.}

\item{save_path}{Path to save the converted \code{obj}.}

\item{quicksave_hdf5}{When \code{filetype="h5"},
\link[HDF5Array]{quickResaveHDF5SummarizedExperiment} will be used to save
the data.}

\item{overwrite}{If a file of the same name exists, overwrite it.}

\item{verbose}{Print messages.}
}
\description{
Save a variety of single-cell data formats.
}
\examples{
obj <- example_obj("cds")
filepath <- save_data(obj)
}
