% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_merge.R
\name{sce_merge}
\alias{sce_merge}
\title{Merge a list of \code{SingleCellExperiment} objects}
\usage{
sce_merge(
  sce_list,
  method = "union",
  cut_off_batch = 0.01,
  cut_off_overall = 0.01,
  use_assays = NULL,
  colData_names = intersect,
  batch_names = names(sce_list),
  verbose = TRUE
)
}
\arguments{
\item{sce_list}{A list contains the \code{SingleCellExperiment}
Object from each batch.}

\item{method}{A string indicates the method of combining the
gene expression matrix, either \code{union} or \code{intersect}.
 Default to \code{intersect}. \code{union} only supports matrix class.}

\item{cut_off_batch}{A numeric vector indicating the cut-off for
the proportion of a gene is expressed within each batch.}

\item{cut_off_overall}{A numeric vector  indicating the cut-off for
the proportion of a gene is expressed overall data.}

\item{use_assays}{A string vector indicating the expression matrices
to be combined.
The first assay named will be used to determine the proportion of zeros.}

\item{colData_names}{A string vector indicating the \code{colData}
that are combined.}

\item{batch_names}{A string vector indicating the batch names for
the output SCE object.}

\item{verbose}{Print messages.}
}
\value{
A \code{SingleCellExperiment} object
with the list of SCE objects combined.
}
\description{
Combine several \code{SingleCellExperiment} objects from
different batches/experiments.
Extracted from the
 \href{https://bioconductor.org/packages/release/bioc/html/scMerge.html}{
 scMerge package}.
}
\examples{
obj <- example_obj("sce")
batch_names<-unique(obj$batch)
sce_list<-list(obj[,obj$batch=='batch2'],
               obj[,obj$batch=='batch3'])
obj2 <- sce_merge(sce_list, batch_names = batch_names)
}
