% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_to_se.R
\name{sce_to_se}
\alias{sce_to_se}
\title{Convert: \code{SingleCellExperiment} ==> \code{SummarizedExperiment}}
\usage{
sce_to_se(obj, verbose = TRUE)
}
\description{
Convert: \code{SingleCellExperiment} ==> \code{SummarizedExperiment}
}
\examples{
obj <- example_obj("sce")
obj2 <- sce_to_se(obj)
}
