% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_to_seurat.R
\name{sce_to_seurat}
\alias{sce_to_seurat}
\title{Convert: \code{SingleCellExperiment} ==> \code{Seurat}}
\usage{
sce_to_seurat(object, verbose = T)
}
\description{
Convert: \code{SingleCellExperiment} ==> \code{Seurat}
}
\examples{
library(scKirby)
data("example_sce")
seurat <- sce_to_seurat(example_sce)
}
