% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{scKirby-package}
\alias{scKirby}
\alias{scKirby-package}
\title{scKirby: Automated ingestion and conversion of various single-cell data formats.}
\description{
Automated ingestion and conversion of various single-cell data formats.
}
\details{
Automated ingestion and conversion of various single-cell data formats.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/neurogenomics/scKirby}
  \item Report bugs at \url{https://github.com/neurogenomics/scKirby/issues}
}

}
\author{
Brian M. Schilder
}
