% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_to_anndata.R
\name{se_to_anndata}
\alias{se_to_anndata}
\title{Convert: \code{SummarizedExperiment} ==> \code{AnnData}}
\usage{
se_to_anndata(obj, verbose = TRUE)
}
\description{
Convert: \code{SummarizedExperiment} ==> \code{AnnData}
}
\examples{
obj <- example_obj("se")
obj2 <- se_to_anndata(obj)
}
