% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_to_le.R
\name{se_to_le}
\alias{se_to_le}
\title{Convert: \code{SummarizedExperiment} ==> \code{SingleCellLoomExperiment}}
\usage{
se_to_le(obj, as_scle = FALSE, verbose = TRUE)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{as_scle}{Convert to \link[LoomExperiment]{LoomExperiment} format.}

\item{verbose}{Print messages.}
}
\value{
A \link[LoomExperiment]{LoomExperiment} object.
}
\description{
Convert: \code{SummarizedExperiment} ==> \code{SingleCellLoomExperiment}
}
\examples{
obj <- example_obj("se")
obj2 <- se_to_le(obj)
}
