% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_to_sce.R
\name{se_to_sce}
\alias{se_to_sce}
\title{Convert: \code{SummarizedExperiment} ==> \code{SingleCellExperiment}}
\usage{
se_to_sce(obj, verbose = TRUE)
}
\description{
Convert: \code{SummarizedExperiment} ==> \code{SingleCellExperiment}
}
\examples{
obj <- example_obj("se")
obj2 <- se_to_sce(obj)
}
