% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_to_scle.R
\name{se_to_scle}
\alias{se_to_scle}
\title{Convert: \code{SummarizedExperiment} ==> \code{SingleCellLoomExperiment}}
\usage{
se_to_scle(obj, verbose = TRUE)
}
\description{
Convert: \code{SummarizedExperiment} ==> \code{SingleCellLoomExperiment}
}
\examples{
obj <- example_obj("se")
obj2 <- se_to_scle(obj)
}
