% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_to_seurat.R
\name{se_to_seurat}
\alias{se_to_seurat}
\title{Convert: \code{SummarizedExperiment} ==> \code{Seurat}}
\usage{
se_to_seurat(obj, verbose = TRUE)
}
\description{
Convert: \code{SummarizedExperiment} ==> \code{Seurat}
}
\examples{
obj <- example_obj("se")
obj2 <- se_to_seurat(obj)
}
