% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_graph.R
\name{set_graph}
\alias{set_graph}
\title{Set graphs}
\usage{
set_graph(obj, g, key, verbose = TRUE)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{g}{A graph.}

\item{key}{Name of the graph key to set \code{g} as within \code{obj}.}

\item{verbose}{Print messages.}
}
\value{
Named list of graph objects.
}
\description{
Set graph objects in any single-cell object class.
}
\examples{
obj <- example_obj("seurat")
g <- get_cor(obj)
obj2 <- set_graph(obj=obj, g=g, key="cor_graph")
}
