% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_uns.R
\name{set_uns}
\alias{set_uns}
\title{Set unstructured data}
\usage{
set_uns(obj, uns, key, verbose = TRUE)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{uns}{Unstructured data to be stored in object.}

\item{key}{Name of the list element to store \code{uns} in.}

\item{verbose}{Print messages.}
}
\value{
Single-cell object.
}
\description{
Set the unstructured data slot in any single-cell object that has one.
}
\examples{
obj <- example_obj("ad")
uns <- list("extra_info"=mtcars)
obj2 <- set_uns(obj = obj,
                uns = mtcars,
                key = "extra_info")
}
