% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_var.R
\name{set_var}
\alias{set_var}
\title{Set unstructured data}
\usage{
set_var(obj, var, verbose = TRUE)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{var}{Variable metadata
as a \link[base]{data.frame} with features as row names.}

\item{verbose}{Print messages.}
}
\value{
Single-cell object.
}
\description{
Set the feature variable slot (i.e. gene metadata)
in any single-cell object that has one.
}
\examples{
obj <- example_obj("seurat")
var <- get_var(obj)
var$new_col <- c(1,2)
obj2 <- set_var(obj = obj,
                var = var)
}
