% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat1_to_list.R
\name{seurat1_to_list}
\alias{seurat1_to_list}
\title{Convert: \code{Seurat (V1)} ==> \code{list}}
\usage{
seurat1_to_list(obj, verbose = TRUE)
}
\description{
Convert: \code{Seurat (V1)} ==> \code{list}
}
