% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_to_anndata.R
\name{seurat_to_anndata}
\alias{seurat_to_anndata}
\title{Convert: \code{Seurat} ==> \code{AnnData}}
\usage{
seurat_to_anndata(
  obj,
  reimport = TRUE,
  save_path = tempfile(fileext = ".h5ad"),
  method = c("sceasy", "anndatar"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{reimport}{Save and re-import the \code{AnnData} object into R to ensure
all data has been converted from Python-native to R-native objects
(e.g. pandas data.frames vs. R data.frames).}

\item{method}{R package to use when converting object.}

\item{...}{
  Arguments passed on to \code{\link[anndata:write_h5ad]{anndata::write_h5ad}}
  \describe{
    \item{\code{anndata}}{An \code{\link[anndata:AnnData]{AnnData()}} object}
    \item{\code{filename}}{Filename of data file. Defaults to backing file.}
    \item{\code{compression}}{See the h5py \href{http://docs.h5py.org/en/latest/high/dataset.html#dataset-compression}{filter pipeline}.
Options are \code{"gzip"}, \code{"lzf"} or \code{NULL}.}
    \item{\code{compression_opts}}{See the h5py \href{http://docs.h5py.org/en/latest/high/dataset.html#dataset-compression}{filter pipeline}.}
    \item{\code{as_dense}}{Sparse in AnnData object to write as dense. Currently only supports \code{"X"} and \code{"raw/X"}.}
  }}
}
\description{
Convert: \code{Seurat} ==> \code{AnnData}
}
\examples{
obj <- example_obj("seurat")
obj2 <- seurat_to_anndata(obj)
}
