% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_to_cds.R
\name{seurat_to_cds}
\alias{seurat_to_cds}
\title{Convert: \code{Seurat} ==> \code{CellDataSet}}
\usage{
seurat_to_cds(obj, verbose = TRUE)
}
\description{
Convert: \code{Seurat} ==> \code{CellDataSet}
}
\examples{
obj <- example_obj("Seurat")
obj2 <- seurat_to_cds(obj)
}
