% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_to_h5seurat.R
\name{seurat_to_h5seurat}
\alias{seurat_to_h5seurat}
\title{Convert: \code{Seurat} ==> \code{H5Seurat}}
\usage{
seurat_to_h5seurat(
  obj,
  save_path = tempfile(fileext = ".h5seurat"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{save_path}{Path to save the \code{H5Seurat} to.}

\item{verbose}{Print messages.}

\item{...}{
  Arguments passed on to \code{\link[SeuratDisk:SaveH5Seurat]{SeuratDisk::as.h5Seurat}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Convert: \code{Seurat} ==> \code{H5Seurat}
}
\examples{
obj <- example_obj("seurat")
h5seurat <- seurat_to_h5seurat(obj)
}
