% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_to_list.R
\name{seurat_to_list}
\alias{seurat_to_list}
\title{Convert: \code{Seurat} ==> \code{list}}
\usage{
seurat_to_list(obj)
}
\description{
Convert: \code{Seurat} ==> \code{list}
}
\examples{
obj <- example_obj("seurat")
obj2 <- seurat_to_list(obj)
}
