% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_to_sce.R
\name{seurat_to_sce}
\alias{seurat_to_sce}
\title{Convert: \code{Seurat} ==> \code{SingleCellExperiment}}
\usage{
seurat_to_sce(object, verbose = T)
}
\description{
Convert: \code{Seurat} ==> \code{SingleCellExperiment}
}
\examples{
library(scKirby)
data("example_seurat")
sce <- seurat_to_sce(example_seurat)
}
