% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_all.R
\name{source_all}
\alias{source_all}
\title{source_all}
\usage{
source_all(path = "R/", pattern = "*.R$", packages = "dplyr")
}
\description{
Source all files in a directory at once. 
Also loads selected libraries.
}
\keyword{internal}
