% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_delayedarray.R
\name{to_delayedarray}
\alias{to_delayedarray}
\title{Convert: to \code{DelayedArray}}
\usage{
to_delayedarray(obj, as_sparse = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{as_sparse}{Convert to a \link[Matrix]{sparseMatrix}.}

\item{verbose}{Print messages.}

\item{...}{
  Arguments passed on to \code{\link[=get_x]{get_x}}
  \describe{
    \item{\code{n}}{The maximum number of elements to return from that list.}
    \item{\code{transpose}}{Transpose the matrix with \code{Matrix::t}.}
    \item{\code{assay}}{Name of the initial assay}
    \item{\code{slot}}{Name of component object to return}
  }}
}
\value{
A \link[DelayedArray]{DelayedArray}.
}
\description{
Convert any object to \link[DelayedArray]{DelayedArray} format.
}
\examples{
obj <- data.frame(gene=paste0("gene",seq_len(10)),
                  matrix(data = 0,nrow=10,ncol = 20))
obj2 <- to_delayedarray(obj)
}
