% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_graph.R
\name{to_graph}
\alias{to_graph}
\title{To graph}
\usage{
to_graph(obj, verbose = TRUE)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{verbose}{Print messages.}
}
\value{
A sparse \link[Seurat]{as.Graph} object.
}
\description{
Convert a matrix to a Graph object (using \link[Seurat]{as.Graph}).
}
\examples{
obj <- example_obj("matrix")
obj2 <- to_graph(obj)
}
