% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_list.R
\name{to_list}
\alias{to_list}
\title{Convert: to \code{list}}
\usage{
to_list(obj, verbose = TRUE)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{verbose}{Print messages.}
}
\value{
A \link[base]{list} object.
}
\description{
Convert any object to \link[base]{list} format.
}
\examples{
obj <- example_obj("seurat")
obj2 <- to_list(obj)
}
