% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_loom.R
\name{to_loom}
\alias{to_loom}
\title{Convert: to \code{loom}}
\usage{
to_loom(obj, save_path = file.path(tempdir(), "scKirby.loom"), verbose = TRUE)
}
\description{
Convert any object to \code{loom} format.
}
\examples{
obj <- example_obj("seurat")
obj2 <- to_loom(obj)
}
