% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_seurat.R
\name{to_seurat}
\alias{to_seurat}
\title{To \code{Seurat}}
\usage{
to_seurat(
  obj,
  as_h5seurat = FALSE,
  save_path = tempfile(fileext = ".h5seurat"),
  verbose = TRUE
)
}
\arguments{
\item{as_h5seurat}{Convert to the \code{H5Seurat} sublass.}

\item{save_path}{Path to save the \code{H5Seurat} to.}
}
\description{
Convert any single-cell object to \code{Seurat} format.
}
\examples{
obj <- example_obj("cds")
obj2 <- to_seurat(obj)
}
