% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_seurat.R
\name{update_seurat}
\alias{update_seurat}
\title{Update Seurat object}
\usage{
update_seurat(obj, verbose = TRUE)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{verbose}{Print messages.}
}
\description{
Update a Seurat from V1 --> V3 or V2 --> V3.
}
\examples{
obj <- example_obj("seurat")
obj2 <- update_seurat(obj)
}
