% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anndata_to_ctd.R
\name{anndata_to_ctd}
\alias{anndata_to_ctd}
\title{Convert: \code{AnnData} ==> \code{CellTypeDataset}}
\usage{
anndata_to_ctd(
  obj,
  annotLevels,
  dataset = basename(tempfile()),
  chunk_size = NULL,
  agg_fun = "mean",
  agg_method = c("monocle3", "stats"),
  dropNA = TRUE,
  standardise = TRUE,
  as_sparse = TRUE,
  as_DelayedArray = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{agg_fun}{Aggregation function.}

\item{agg_method}{Aggregation method.}

\item{dropNA}{Drop genes assigned to \code{NA} in \code{groupings}.}

\item{as_sparse}{Convert aggregated matrix to sparse matrix.}

\item{as_DelayedArray}{Convert aggregated matrix to
\link[DelayedArray]{DelayedArray}.}

\item{verbose}{Print messages.}
}
\description{
Convert: \code{AnnData} ==> \code{CellTypeDataset}
}
\examples{
obj <- example_obj("anndata")
obj2 <- anndata_to_ctd(obj)
}
