% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anndata_to_list.R
\name{anndata_to_list}
\alias{anndata_to_list}
\title{Convert: \code{anndata} ==> \code{list}}
\usage{
anndata_to_list(obj)
}
\description{
Convert: \code{anndata} ==> \code{list}
}
\examples{
obj <- example_obj("anndata")
obj2 <- anndata_to_list(obj)
}
