% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anndata_to_se.R
\name{anndata_to_se}
\alias{anndata_to_se}
\title{Convert: \code{Anndata} ==> \code{SummarizedExperiment}}
\usage{
anndata_to_se(obj, as_sce = FALSE, verbose = TRUE)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{as_sce}{Convert to a \link[SingleCellExperiment]{SingleCellExperiment}
format.}

\item{verbose}{Print messages.}
}
\description{
Convert: \code{Anndata} ==> \code{SummarizedExperiment}
}
\examples{
obj <- example_obj("anndata")
sce <- anndata_to_se(obj)
}
