% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anndata_to_seurat.R
\name{anndata_to_seurat}
\alias{anndata_to_seurat}
\title{Convert: \code{Anndata} ==> \code{Seurat}}
\usage{
anndata_to_seurat(object, save_dir = tempdir(), verbose = T)
}
\description{
Convert: \code{Anndata} ==> \code{Seurat}
}
\examples{
library(scKirby)
seurat <- anndata_to_seurat(object=example_anndata())
}
