% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_to_sce.R
\name{cds_to_se}
\alias{cds_to_se}
\title{Convert: \code{CellDataSet} ==> \code{SummarizedExperiment}}
\usage{
cds_to_se(
  obj,
  as_sce = FALSE,
  as_sparse = TRUE,
  as_DelayedArray = FALSE,
  verbose = TRUE
)
}
\description{
Convert: \code{CellDataSet} ==> \code{SummarizedExperiment}
}
\examples{
obj <- example_obj("cds")
obj2 <- cds_to_se(obj)
}
