% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_to_seurat.R
\name{cds_to_seurat}
\alias{cds_to_seurat}
\title{Convert: \code{CellDataSet} ==> \code{Seurat}}
\usage{
cds_to_seurat(obj, verbose = TRUE)
}
\description{
Convert: \code{CellDataSet} ==> \code{Seurat}
}
\examples{
obj <- example_obj("cds")
obj2 <- cds_to_seurat(obj)
}
