% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd_to_se.R
\name{ctd_to_se}
\alias{ctd_to_se}
\title{Convert: \code{CellTypeDataset} ==> \code{SingleCellExperiment}}
\usage{
ctd_to_se(obj, as_sce = FALSE, verbose = TRUE)
}
\description{
Convert: \code{CellTypeDataset} ==> \code{SingleCellExperiment}
}
\examples{
obj <- example_obj("ctd")
obj <- ctd_to_sce(example_ctd)
}
