% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd_to_seurat.R
\name{ctd_to_seurat}
\alias{ctd_to_seurat}
\title{Convert: \code{CellTypeDataSet} ==> \code{Seurat}}
\usage{
ctd_to_seurat(obj, verbose = TRUE)
}
\description{
Convert: \code{CellTypeDataSet} ==> \code{Seurat}
}
\examples{
data("example_ctd")
obj2 <- ctd_to_seurat(example_ctd)
}
