% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_anndata.R
\name{example_anndata}
\alias{example_anndata}
\title{Example \code{anndata}}
\source{
\code{
#### Setup conda env ####
## Option 1: Point to where anndata is installed (or should be installed)
conda_dir <- dirname(dirname(reticulate::conda_list()[1,]$python))
reticulate::use_condaenv(condaenv = conda_dir)
reticulate::conda_install(conda = conda_dir, packages = "loompy", pip = T)
## Option 2: Just run anndata::install_anndata() and will install via miniconda
anndata::install_anndata(method = "conda", conda=conda_dir)
#### Usage ####
adata <- example_anndata()
}
}
\usage{
example_anndata(
  obj = SeuratObject::pbmc_small,
  save_path = file.path(tempdir(), "example.h5ad"),
  verbose = TRUE
)
}
\description{
\href{https://github.com/dynverse/anndata/issues/1}{
anndata R package has a major bug currently.}
Thus, \code{add_filename} must be kept \code{False} until this is fixed.
}
\keyword{internal}
