% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_obj.R
\name{example_obj}
\alias{example_obj}
\title{Example single-cell data object}
\usage{
example_obj(
  class = c("SummarizedExperiment", "SingleCellExperiment", "Seurat", "h5Seurat",
    "CellDataSet", "matrix", "data.table", "data.frame", "loom", "LoomExperiment",
    "SingleCellLoomExperiment", "anndata", "EWCE"),
  save_path = file.path(tempdir(), "example"),
  verbose = TRUE
)
}
\value{
Single-cell object.
}
\description{
Create a small example object
of scRNA-seq data.
}
\details{
\emph{NOTE:} If you get the error
 \code{Error: ModuleNotFoundError: No module named 'anndata'},
 try running \link[anndata]{install_anndata} and
 then restarting your R session.

Certain file types (loom, anndata) can't be stored a .rda
 because they need to be set up for on-disk reading.
}
\examples{
se <- example_obj("se")
sce <- example_obj("sce")
seurat <- example_obj("seurat")

library(Seurat) # necessary
loom <- example_obj("loom")
}
