% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_graphs.R
\name{get_graphs}
\alias{get_graphs}
\title{Get graphs}
\usage{
get_graphs(obj, names = NULL, verbose = TRUE)
}
\arguments{
\item{names}{The names of reductions to extract from.}

\item{verbose}{Print messages.}
}
\description{
Get graph objects from any single-cell object class.
}
\examples{
obj <- example_obj("seurat")
g <- get_graphs(obj)
}
