% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_anndata.R
\name{list_to_anndata}
\alias{list_to_anndata}
\title{Convert: \code{list} ==> \code{AnnData}}
\usage{
list_to_anndata(obj, transpose = TRUE, verbose = TRUE)
}
\description{
Convert: \code{list} ==> \code{AnnData}
}
\examples{
obj <- example_obj("list")
obj2 <- list_to_anndata(obj)
}
