% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_seurat.R
\name{list_to_seurat}
\alias{list_to_seurat}
\title{Convert: \code{matrix} ==> \code{Seurat}}
\usage{
list_to_seurat(obj, verbose = TRUE)
}
\description{
Convert: \code{matrix} ==> \code{Seurat}
}
\examples{
obj <- example_obj("list")
obj2 <- list_to_seurat(obj)
}
