% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rdata.R
\name{load_rdata}
\alias{load_rdata}
\title{load RData}
\usage{
load_rdata(fileName)
}
\description{
Load processed data using a function that assigns it
to a specific variable (so you don't have to guess what the loaded variable name is).
}
