% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom_to_scle.R
\name{loom_to_scle}
\alias{loom_to_scle}
\title{Convert: \code{loom} ==> \code{SingleCellLoomExperiment}}
\usage{
loom_to_scle(obj, verbose = TRUE, ...)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{verbose}{Print messages.}

\item{...}{
    For constructors, \code{...} will be passed on to the respective Experiment constructor.
  }
}
\description{
Convert: \code{loom} ==> \code{SingleCellLoomExperiment}
}
\examples{
library(Seurat)
obj <-  example_obj("loom")
sce <- loom_to_scle(obj)
}
