% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom_to_seurat.R
\name{loom_to_seurat}
\alias{loom_to_seurat}
\title{Convert: \code{loom} ==> \code{Seurat}}
\usage{
loom_to_seurat(object, verbose = T)
}
\description{
Convert: \code{loom} ==> \code{Seurat}
}
\examples{
library(scKirby)
loom <- example_loom()
seurat <- loom_to_seurat(loom)
}
