% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom_to_seurat.R
\name{loom_to_seurat}
\alias{loom_to_seurat}
\title{Convert: \code{loom} ==> \code{Seurat}}
\usage{
loom_to_seurat(obj, verbose = TRUE)
}
\description{
Convert: \code{loom} ==> \code{Seurat}
}
\examples{
library(Seurat)
obj <- example_obj("loom")
obj2 <- loom_to_seurat(obj)
}
