% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_se.R
\name{matrix_to_se}
\alias{matrix_to_se}
\title{Convert: \code{matrix} ==> \code{SummarizedExperiment}}
\usage{
matrix_to_se(obj, as_sce = FALSE, verbose = TRUE)
}
\description{
Convert: \code{matrix} ==> \code{SummarizedExperiment}
}
\examples{
obj <- example_obj("matrix")
obj2 <- matrix_to_se(obj)
}
