% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_seurat.R
\name{matrix_to_seurat}
\alias{matrix_to_seurat}
\title{Convert: \code{matrix} ==> \code{Seurat}}
\usage{
matrix_to_seurat(object, verbose = T)
}
\description{
Convert: \code{matrix} ==> \code{Seurat}
}
\examples{
library(scKirby)
data("example_seurat")
sce <- matrix_to_seurat(object=example_seurat@assays$RNA@counts)
}
