% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_seurat.R
\name{merge_seurat}
\alias{merge_seurat}
\title{Merge \link[Seurat]{Seurat} objects}
\usage{
merge_seurat(obj_list, batch_size = 5, verbose = TRUE)
}
\arguments{
\item{obj_list}{A named list of \link[Seurat]{Seurat} objects,
or paths to \link[Seurat]{Seurat} objects saved to disk as \emph{.rds} files.}
}
\value{
A merged \link[Seurat]{Seurat} object.
}
\description{
Merge a list of Seurat objects efficiently.
}
\examples{
obj <- example_obj("seurat")
obj_list <- Seurat::SplitObject(obj,split.by = "groups")
obj2 <- merge_seurat(obj_list)
}
