% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_seurat.R
\name{process_seurat}
\alias{process_seurat}
\title{Process a \link[SeuratObject]{Seurat} object}
\usage{
process_seurat(
  obj = NULL,
  meta.data = NULL,
  nfeatures = 2000,
  vars.to.regress = NULL,
  dims = seq_len(50),
  add_specificity = FALSE,
  assay_name = "RNA",
  default_assay = NULL,
  n.components = 2L,
  log_norm = FALSE,
  cluster_reduction = "umap",
  workers = 1,
  max_mem = 8000 * 1024^2,
  seed = 2020
)
}
\arguments{
\item{cluster_reduction}{Recompute neighbors graph based on UMAP
to get clusters that best reflect UMAP space.
For this same reason, only cluster in two dimensions,
because this is the view we most often use.
That said, this may reduce the generalisability of these clusters/graph.}
}
\value{
A preprocessed \link[Seurat]{Seurat} object.
}
\description{
Run a standardised \link[SeuratObject]{Seurat} pipeline on
a \link[SeuratObject]{Seurat} object
or raw \code{counts} and \code{meta.data}.\cr
Automatically performs
\describe{
\item{\code{FindVariableFeatures}}{Variable feature selection}
\item{\code{NormalizeData}}{Data normalization}
\item{\code{RunPCA}}{PCA}
\item{\code{RunUMAP}}{UMAP}
\item{\code{FindNeighbors}}{K-nearest neighbors}
\item{\code{FindClusters}}{Clustering}
}
}
\examples{
obj <- example_obj("matrix")
obj2 <- seurat_pipeline(obj = obj)
}
