% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data.R
\name{save_data}
\alias{save_data}
\title{Save data}
\usage{
save_data(
  obj,
  filetype = c("h5", "h5seurat", "h5ad", "rda", "rds"),
  save_path = file.path(tempdir(), paste("scKirby_output", filetype, sep = ".")),
  quicksave_HDF5 = TRUE,
  overwrite = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{obj}{A single-cell data object, or a path to saved single-cell data.}

\item{overwrite}{If a file of the same name exists, overwrite it.}
}
\description{
Save a variety of single-cell data formats.
}
\examples{
obj <- example_obj("cds")
filepath <- save_data(obj)
}
