% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_to_se.R
\name{sce_to_se}
\alias{sce_to_se}
\title{Convert: \code{SingleCellExperiment} ==> \code{SummarizedExperiment}}
\usage{
sce_to_se(obj, verbose = TRUE)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{verbose}{Print messages.}
}
\description{
Convert: \code{SingleCellExperiment} ==> \code{SummarizedExperiment}
}
\examples{
obj <- example_obj("sce")
obj2 <- sce_to_se(obj)
}
