% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scle_to_le.R
\name{scle_to_le}
\alias{scle_to_le}
\title{Convert: \code{SingleCellLoomExperiment} ==> \code{LoomExperiment}}
\usage{
scle_to_le(obj, verbose = TRUE)
}
\description{
Convert: \code{SingleCellLoomExperiment} ==> \code{LoomExperiment}
}
\examples{
obj <- example_obj("se")
obj2 <- se_to_le(obj)
}
