% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_to_anndata.R
\name{se_to_anndata}
\alias{se_to_anndata}
\title{Convert: \code{SummarizedExperiment} ==> \code{AnnData}}
\usage{
se_to_anndata(
  obj,
  method = c("zellkonverter", "sceasy"),
  reimport = TRUE,
  save_path = tempfile(fileext = ".h5ad"),
  verbose = TRUE
)
}
\arguments{
\item{method}{Which R package to use for the conversion.}

\item{reimport}{Save and re-import the \code{anndata} object into R to ensure
all data has been converted from Python-native to R-native objects
(e.g. pandas data.frames vs. R data.frames).}

\item{verbose}{Print messages.}
}
\description{
Convert: \code{SummarizedExperiment} ==> \code{AnnData}
}
\examples{
obj <- example_obj("se")
obj2 <- se_to_anndata(obj)
}
