% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_to_list.R
\name{se_to_list}
\alias{se_to_list}
\title{Convert: \code{SummarizedExperiment} ==> \code{list}}
\usage{
se_to_list(obj)
}
\description{
Convert: \code{SummarizedExperiment} ==> \code{list}
}
\examples{
obj <- example_obj("se")
obj2 <- se_to_list(obj)
}
