% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_to_loom.R
\name{se_to_loom}
\alias{se_to_loom}
\title{Convert: \code{SummarizedExperiment} ==> \code{loom}}
\usage{
se_to_loom(
  obj,
  save_path = file.path(tempdir(), "scKirby.loom"),
  verbose = TRUE
)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{save_path}{Path to save the \code{H5Seurat} to.}

\item{verbose}{Print messages.}
}
\description{
Convert: \code{SummarizedExperiment} ==> \code{loom}
}
\examples{
obj <- example_obj("se")
obj2 <- se_to_loom(obj)
}
