% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_obs.R
\name{set_obs}
\alias{set_obs}
\title{Set unstructured data}
\usage{
set_obs(obj, obs, verbose = TRUE)
}
\arguments{
\item{obs}{Observation metadata
as a \link[base]{data.frame} with samples as row names.}
}
\value{
Single-cell object.
}
\description{
Set the sample observations slot (i.e. cell metadata)
 in any single-cell object that has one.
}
\examples{
obj <- example_obj("ad")
obs <- get_obs(obj)
obs$new_col <- c(1,2)
obj2 <- set_obs(obj = obj,
                obs = obs)
}
