% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat1_to_list.R
\name{seurat1_to_list}
\alias{seurat1_to_list}
\title{Convert: \code{Seurat (V1)} ==> \code{list}}
\usage{
seurat1_to_list(obj, verbose = TRUE)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{verbose}{Print messages.}
}
\description{
Convert: \code{Seurat (V1)} ==> \code{list}
}
