% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_to_anndata.R
\name{seurat_to_anndata}
\alias{seurat_to_anndata}
\title{Convert: \code{Seurat} ==> \code{anndata}}
\usage{
seurat_to_anndata(
  obj,
  reimport = TRUE,
  save_path = tempfile(fileext = ".h5ad"),
  verbose = TRUE
)
}
\arguments{
\item{reimport}{Save and re-import the \code{anndata} object into R to ensure
all data has been converted from Python-native to R-native objects
(e.g. pandas data.frames vs. R data.frames).}
}
\description{
Convert: \code{Seurat} ==> \code{anndata}
}
\examples{
obj <- example_obj("seurat")
obj2 <- seurat_to_anndata(obj)
}
