% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_to_h5seurat.R
\name{seurat_to_h5seurat}
\alias{seurat_to_h5seurat}
\title{Convert: \code{Seurat} ==> \code{H5Seurat}}
\usage{
seurat_to_h5seurat(
  obj,
  save_path = tempfile(fileext = ".h5seurat"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{save_path}{Path to save the \code{H5Seurat} to.}
}
\description{
Convert: \code{Seurat} ==> \code{H5Seurat}
}
\examples{
obj <- example_obj("seurat")
seurat <- seurat_to_h5seurat(obj)
}
