% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_to_scle.R
\name{seurat_to_le}
\alias{seurat_to_le}
\title{Convert: \code{Seurat} ==> \code{LoomExperiment}}
\usage{
seurat_to_le(obj, as_scle = FALSE, verbose = TRUE)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{as_scle}{Convert to \link[LoomExperiment]{LoomExperiment} format.}

\item{verbose}{Print messages.}
}
\description{
Convert: \code{Seurat} ==> \code{LoomExperiment}
}
\examples{
obj <- example_obj("seurat")
obj2 <- seurat_to_le(obj)
}
