% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_to_list.R
\name{seurat_to_list}
\alias{seurat_to_list}
\title{Convert: \code{Seurat} ==> \code{list}}
\usage{
seurat_to_list(obj, verbose = TRUE)
}
\arguments{
\item{obj}{A single-cell object supported by \pkg{scKirby}.
See \link[scKirby]{converters} for a table of all supported conversions.}

\item{verbose}{Print messages.}
}
\value{
A named list.
}
\description{
Convert: \code{Seurat} ==> \code{list}
}
\examples{
obj <- example_obj("seurat")
obj2 <- seurat_to_list(obj)
}
