% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_to_se.R
\name{seurat_to_se}
\alias{seurat_to_se}
\title{Convert: \code{Seurat} ==> \code{SummarizedExperiment}}
\usage{
seurat_to_se(obj, as_sce = TRUE, verbose = TRUE)
}
\description{
Convert: \code{Seurat} ==> \code{SummarizedExperiment}
}
\examples{
obj <- example_obj("seurat")
obj2 <- seurat_to_se(obj)
}
