% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_anndata.R
\name{to_anndata}
\alias{to_anndata}
\title{To \code{AnnData}}
\usage{
to_anndata(
  obj,
  save_path = tempfile(fileext = ".h5ad"),
  reimport = FALSE,
  verbose = TRUE
)
}
\description{
Convert any single-cell object to \code{AnnData} format.
}
\examples{
obj <- example_obj("cds")
obj2 <- to_anndata(obj)
}
