% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_list.R
\name{to_list}
\alias{to_list}
\title{Convert: to \code{list}}
\usage{
to_list(obj, verbose = TRUE)
}
\description{
Convert any object to \code{list} format.
}
\examples{
obj <- example_obj("seurat")
l <- to_list(obj)
}
