% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_se.R
\name{to_se}
\alias{to_se}
\title{Convert: ==> \code{SummarizedExperiment}}
\usage{
to_se(obj, as_sce = FALSE, verbose = TRUE)
}
\description{
Convert any object to \code{SummarizedExperiment} or
\code{SingleCellExperiment} format.
}
\examples{
obj <- example_obj("seurat")
obj2 <- to_se(obj)
}
