% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_sparse.R
\name{to_sparse}
\alias{to_sparse}
\title{To sparse matrix}
\usage{
to_sparse(obj, transpose = FALSE, allow_delayed_array = TRUE, verbose = TRUE)
}
\value{
Sparse matrix
}
\description{
Convert a matrix or data.frame to a sparse matrix.
}
\examples{
obj <- data.frame(gene=paste0("gene",seq_len(10)),
                  matrix(data = 0,nrow=10,ncol = 20))
obj2 <- to_sparse(obj)
}
